Update patch set 5
Patch Set 5:

If detule was able to make d2 camera binaries work on the CAF kernel for your device, it should be a fairly similar process to get the mako HAL working and it would certainly be beneficial.  It is certainly not ideal to have to be using binaries.  The HAL is not hardware specific, so it should work on most QCOM devices assuming the correct drivers are in place.

Patch-set: 5