Update patch set 5
Patch Set 5:

Yeah, this can be made to work on a bunch of other qcom hardware. I'm not sure the feature and performance losses are worth it, though; 
- most devices' liboemcamera and chromatix libraries tend to have hardware-specific optimizations. if you're forced to use another devices' libraries, chances are you'll lose those.
- qualcomm's mainline HALs are far richer in features than AOSP's lightweight version. Not that we use those features (besides ZSL in a few devices)
- The AOSP HAL has some annoying bugs, and since there's no other upstream besides Google to draw patches from we either fix them (and this stuff is tricky :) ) or put up with them.

This discussion is besides this particular patch, though; if maintainers want to switch to open-source variants, more power to them, but do not work under the assumption that this is functionally equivalent to what shipped on the devices; it's not. We want as much open-source as possible, but not if that means compromising the final quality :)

Patch-set: 5
Reviewer: Ricardo Cerqueira <10@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0