Update patch set 5
Patch Set 5:

You may want to compare them before deciding on which one to use, but yeah, the differences are sometimes surprising.
The Optimus G is a good example: the shipping fast-AF and auto exposure compensation are WAY better than source-built.  I'm occasionally tempted to put them on the mako builds, since it's basically the same hw :)

Patch-set: 5